+++ src/model/lang.ml
@@ -212,6 +212,11 @@ end =
 
 (** {3 Type Declarations} *)
 
+and Modalities : sig
+  type t = string list
+end =
+  Modalities
+
 and TypeDecl : sig
   module Field : sig
     type t = {
@@ -219,6 +224,7 @@ and TypeDecl : sig
       doc : Comment.docs;
       mutable_ : bool;
       type_ : TypeExpr.t;
+      modalities : Modalities.t;
     }
   end
 
@@ -232,7 +238,9 @@ and TypeDecl : sig
   end
 
   module Constructor : sig
-    type argument = Tuple of TypeExpr.t list | Record of Field.t list
+    type argument =
+      | Tuple of (TypeExpr.t * Modalities.t) list
+      | Record of Field.t list
 
     type t = {
       id : Identifier.Constructor.t;
@@ -340,6 +348,7 @@ and Value : sig
     doc : Comment.docs;
     type_ : TypeExpr.t;
     ext_attrs : attr list;
+    modalities : Modalities.t;
   }
 end =
   Value
@@ -437,7 +446,7 @@ and Kind : sig
     | Default
     | Abbreviation of Fragment.Type.t
     | Mod of t * string list
-    | With of t * TypeExpr.t * string list
+    | With of t * TypeExpr.t * Modalities.t
     | Kind_of of TypeExpr.t
     | Product of t list
 end =