+++ CHANGES.md
@@ -3,6 +3,7 @@
 - Support for OxCaml unboxed named types (@art-w, #1407)
 - Support for OxCaml zero alloc definitions (@Leonidas-from-XIV, #1422, #1444)
 - Remove requirement for ppx_expect in tests (@jonludlam, #1445)
+- Support for OxCaml modalities (@art-w, #1420)
 
 # 3.2.1